Move creating the form to its own directory.
